requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
tzlocal==2.0.0
urllib3==1.25.8
